pyproject.toml
[project]
name = "stop-save-pchurn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-dtypes>=1.5.1",
    "google-cloud-bigquery>=3.41.0",
    "pandas>=2.3.3",
    "pyarrow>=24.0.0",
]
